+++ screens/ComplimentChat/index.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import {Button} from 'react-native';
-// import {GiftedChat} from 'react-native-gifted-chat';
 import API from '../../utils/API';
 import logo from '../../assets/images/check/check3.png';
-
 import bg from '../../assets/images/check/check3.png'
 import ComplimentCard from "../../components/ComplimentCard"
 import Modal from "react-native-modal";
@@ -21,61 +19,7 @@ import {
   Easing,
 } from 'react-native';
 
-
-
-const win = Dimensions.get('window');
-const styles = StyleSheet.create({
-  
-  view: {
-    flexDirection: 'column',
-    flex: 1,
-    justifyContent: 'space-evenly',
-    alignItems: 'center',
-    backgroundColor: '#F46DCE30',
-  },
-
-  upsyImg: {
-    height: 300,
-    justifyContent: 'center',
-  },
-
-
-  subtitleText: {
-    fontSize: 20,
-    textAlign: 'center',
-    color: '#F46DCE',
-    marginTop: 100,
-    padding: 10,
-  },
-
-  text: {
-    fontSize: 30,
-    color: '#F46DCE',
-    textAlign: 'center',
-    flexDirection: 'column',
-    justifyContent: 'center'
-  },
-
-  background: {
-    height: 50,
-    backgroundColor: 'white',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: 350,
-    marginBottom: 250,
-    borderRadius: 15,
-  },
-
-  modalSpacing: {
-    flexDirection: 'column',
-    flex: 1,
-    justifyContent: 'space-around',
-    padding: 15
-  }
-
-  
-});
+import styles from './styles.js'
 
 
 export default class ComplimentChat extends React.Component {
@@ -86,7 +30,6 @@ export default class ComplimentChat extends React.Component {
       compliment: "Click on Upsy for a compliment",
       animationStarted: false,
       // hideUpsy: false,
-
       isModalVisible: false
     };
     this.getUpsyCompliment = this.getUpsyCompliment.bind(this)
@@ -115,11 +58,9 @@ export default class ComplimentChat extends React.Component {
             _id: 2,
             name: 'Upsy',
           },
-          // Any additional custom parameters are passed through
         };
         console.log(upsyMessage);
         
-        //Append Upsy's message to the chat
         this.setState({
           compliment: upsyMessage.text,
         });
@@ -153,59 +94,56 @@ render() {
     modal = <ComplimentCard compliment={this.state.compliment} />
   }
 
-  return (
-    <View
-      style={ styles.view }>
+    return (
+      <View
+        style={ styles.view }>
 
-      <View >
-        <Text style={styles.subtitleText} onPress={() => this.showCompliment()}>Press on Upsy to get started</Text>
+        <View >
+          <Text style={styles.subtitleText} onPress={() => this.showCompliment()}>Press on Upsy to get started</Text>
+
+        </View>
+        <View >
+          <TouchableOpacity
+            onPress={() => {
+              this.getUpsyCompliment();
+              this.toggleModal();
+            }}>
+
+            <Image
+              style={ styles.upsyImg }
+              source={require("../../assets/images/upsy_emo/upsy1_emo8.png")}
+              resizeMode={'contain'}
+            />
+
+          </TouchableOpacity>
+        </View>
+        
+        {/* MODAL START */}
+        <View >
+          <Modal 
+            style={ styles.background }
+
+              isVisible={this.state.isModalVisible}
+
+              swipeDirection='right'
+              animationIn='slideInUp'
+              animationInTiming={500}
+              animationOut='slideOutDown'
+              animationOutTiming={750}
+              backdropColor='#fff'
+              backdropOpacity={.5}
+
+            >
+            <View  style={ styles.modalSpacing } >
+              <Text style={styles.text} >
+                  {this.state.compliment}
+              </Text>
+              <Button title="Done" onPress={this.toggleModal} />
+            </View>
+          </Modal>
+        </View>
 
       </View>
-      <View >
-        <TouchableOpacity
-          onPress={() => {
-            this.getUpsyCompliment();
-            this.toggleModal();
-          }}>
-
-          <Image
-            style={ styles.upsyImg }
-            source={require("../../assets/images/upsy_emo/upsy1_emo8.png")}
-            resizeMode={'contain'}
-          />
-
-        </TouchableOpacity>
-      </View>
-      
-      {/* MODAL START */}
-      <View >
-        <Modal 
-        style={ styles.background }
-
-            isVisible={this.state.isModalVisible}
-
-            swipeDirection='right'
-            animationIn='slideInUp'
-            animationInTiming={500}
-            animationOut='slideOutDown'
-            animationOutTiming={750}
-            backdropColor='#fff'
-            backdropOpacity={.5}
-
-        >
-          <View  style={ styles.modalSpacing } >
-            <Text style={styles.text} >
-                {this.state.compliment}
-            </Text>
-            <Button title="Done" onPress={this.toggleModal} />
-          </View>
-        </Modal>
-      </View>
-
-    </View>
-  )
-}
-
-
-
+    )
+  }
 }
